Change all instances of floats to doubles for increasing precision and making serialization easier (no need to chop off float errors)

+++ src/Defines.h
@@ -8,6 +8,6 @@
 #define WINDOW_FLAGS SDL_WINDOW_SHOWN //SDL_WINDOW_RESIZABLE | SDL_WINDOW_SHOWN
 
 constexpr unsigned int FPS = 60;
-constexpr float SECOND_CHANGE_PER_FRAME = (1.0f / (float)FPS);
+constexpr double SECOND_CHANGE_PER_FRAME = 1.0 / FPS;
 
 #define DEFAULT_RENDER_COLOR SDL_Color{ 255, 255, 255, 255 }
