+++ src/TextureManager.h
@@ -13,7 +13,7 @@ public:
 	static SDL_Texture* load(std::string path);
 	static SDL_Texture* getTexture(const std::string& path);
 	static void draw(SDL_Texture* texture, SDL_Rect source, SDL_Rect destination);
-	static void draw(SDL_Texture* texture, SDL_Rect source, SDL_Rect destination, float angle, SDL_RendererFlip flip);
+	static void draw(SDL_Texture* texture, SDL_Rect source, SDL_Rect destination, double angle, SDL_RendererFlip flip);
 	static void draw(SDL_Rect rectangle, SDL_Color color = { 0, 0, 0, 255 });
 	static void removeTexture(std::string path);
 private:
